+++ compare-frontend/src/app/modules/map/converters/eqts/2_eqsim.ts
@@ -28,6 +28,7 @@ export class EqSimulation implements Converter {
 
             return of([
                 {
+                    id: "eqSimWmsLayer",
                     layer: new TileLayer({
                         source: new TileWMS({
                             url: baseUrl,