+++ CHANGELOG.md
@@ -27,6 +27,7 @@
     - using rules to decide what exactly to do in reducers
     - user can chose between rule-sets
     - can now close popups
+    - popups get features at click location as input
     - Ongoing
         - map has click-behavior
            - add popup to map on click